Using Native calls instead of CopyFromScreen (#232)
Using Native calls instead of `System.Drawing.Graphics.CopyFromScreen`.

Device Contexts and Native Bitmaps are created only once. This takes only `6ms`.
Image Pool cannot be used with this approach.
Bitmap `CopyTo` Buffer function uses `Parallel.For` and accounts for negative stride.

It is not certain whether this is really an improvement, since we are allocating a new `System.Drawing.Bitmap` for every frame.

In `WindowProvider`, `StretchBlt` is used. This seems to be a good performance improvement compared to the current method.

Also, fixed the Recorder handling `OperationCancelledException`s like when a window is closed during recording.